backend/requirements.txt
azure-functions==1.0.4
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
idna==2.8
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.3
pyparsing==2.4.2
python-dateutil==2.8.0
requests==2.22.0
six==1.12.0
urllib3==1.25.6
